moved fetching sectors to related step
